Update some controllers Login JWT system added
TODO: Rest of controllers, JWT Registration

+++ CRM.WebAPI/CRM.WebAPI/Controllers/ClientController.cs
@@ -1,12 +1,15 @@
 using Microsoft.AspNetCore.Mvc;
 using CRM.Domain.Models;
 using CRM.Infrastructure.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 
 namespace CRM.WebAPI.Controllers
 {
     [Route("api/[controller]/[action]")]
+
     //Контроллер - управляет таблицей в данном случае таблицей клиентов
     [ApiController]
+    [Authorize]
     public class ClientController : ControllerBase
     {
         private IClientRepository _repository; //храним репозиторий (что то что напрямую взаимодействует с базой)
@@ -30,7 +33,15 @@ namespace CRM.WebAPI.Controllers
         [HttpGet("{id}")] // ищет по id
         public async Task<ActionResult> GetById(Guid id)
         {
-            return Ok(await _repository.GetById(id));
+            if (id == Guid.Empty)
+                return BadRequest("Id is empty");
+            var result = await _repository.GetById(id);
+            if (result.Successful)
+            {
+                return Ok(result);
+            }
+
+            return NotFound($"Client with id {id} does not exist");
         }
 
 
@@ -39,15 +50,29 @@ namespace CRM.WebAPI.Controllers
         public async Task<IActionResult> Insert([FromBody] Client client)
 
         {
-            await _repository.Put(client);
-            return Ok();
+            if (client.ContactId == Guid.Empty)
+            {
+                return BadRequest("ContactId is empty");
+            }
+
+            var result = await _repository.Put(client);
+            if (result.Successful)
+                return Ok(result);
+            return BadRequest(result.ErrorMessage);
         }
 
         [HttpPost]
         public async Task<ActionResult>
             Update([FromBody] Client dataToUpdate) //обновляет данные (что угодно можно поменять в клиенте кроме id)
         {
-            await _repository.Update(dataToUpdate);
+            if (dataToUpdate.ClientId == Guid.Empty)
+                return BadRequest("Client id is empty!");
+            var result = await _repository.Update(dataToUpdate);
+            if (!result.Successful)
+            {
+                return BadRequest(result.ErrorMessage);
+            }
+
             return Ok();
         }
 
@@ -55,7 +80,17 @@ namespace CRM.WebAPI.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(Guid id) //удаляет по id
         {
-            await _repository.RemoveById(id);
+            if (id == Guid.Empty)
+            {
+                return BadRequest("Client id is empty!");
+            }
+
+            var result = await _repository.RemoveById(id);
+            if (!result.Successful)
+            {
+                return BadRequest(result.ErrorMessage);
+            }
+
             return Ok();
         }
     }
